[Util] Fix expression ambiguity in `Pooled::count()`
The original ternary expression could cause ambiguity since
the return types of its branches could cyclically be converted to
each other. More precisely, if `Pool` is thread-safe, `ref_->second`
will have the type 'std::atomic<unsigned int>' and it could be implicitly
cast to 'int' (return type of 2nd branch) and vice versa.
Fixed by explicitly casting `ref_->second` to `uint32_t`.

+++ include/mutable/util/Pool.hpp
@@ -236,7 +236,7 @@ struct Pooled
      * Returns the number of references to the pooled object or 0 if
      * this `Pooled` CanBeNone and does *not* hold a reference to an object.
      */
-    uint32_t count() const { return ref_ ? ref_->second : 0; }
+    uint32_t count() const { return ref_ ? uint32_t(ref_->second) : 0; }
 
     ~Pooled() {
         M_insist(bool(pool_) == bool(ref_), "inconsistent pooled state");
